+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## cockpit-45drives-hardware 2.0.1-4
+## cockpit-45drives-hardware 2.0.1-5
 
-* updated unsupported motherboard message in 45Drives-motherboard module
+* added ipmi lan info section to 45Drives-system module